requirements.txt
ultralytics
opencv-python
python-dotenv
deep-sort-realtime
